+++ spec/argumental_preload_options_spec.rb
@@ -17,76 +17,75 @@ describe Argumental::Action do
         it 'should have no default options set' do
             subject.args = []
             subject.run
-            subject.info[:top].should == 'nothing'
+            subject.config[:top].should == 'nothing'
         end
 
         it 'should have default options set with strings as keys' do
             subject.args = []
-            subject.set_default_options({'top' => "A String Key"})
+            subject.apply_config({'top' => "A String Key"})
             subject.run
-            subject.info[:top].should == 'A String Key'
+            subject.config[:top].should == 'A String Key'
         end
 
         it 'should have default options set with symbols as keys' do
             subject.args = []
-            subject.set_default_options({top: 'A Symbol Key'})
+            subject.apply_config({top: 'A Symbol Key'})
             subject.run
-            subject.info[:top].should == 'A Symbol Key'
+            subject.config[:top].should == 'A Symbol Key'
         end
 
         it 'should have default options overridden when passed in on commandline' do
             subject.args = ['--top', 'Command line magic']
-            subject.set_default_options({top: 'A Symbol Key'})
+            subject.apply_config({top: 'A Symbol Key'})
             subject.run
-            subject.info[:top].should == 'Command line magic'
+            subject.config[:top].should == 'Command line magic'
         end
 
         it 'should have dynamic preset use depending on option input' do
             subject.args = ['--top', 'Command line magic']
             subject.set_pre_validate do
+                opts = options
                 if options[:top]
-                    set_default_options({sub1: 'Dynamically assigned'})
+                    apply_config({sub1: 'Dynamically assigned'})
                 end
             end
             subject.run
-            subject.info[:top].should == 'Command line magic'
-            subject.info[:sub1].should == 'Dynamically assigned'
-        end
-
-        it 'should have dynamic not preset since dependant option missing' do
-            subject.args = []
-            subject.set_pre_validate do
-                if options[:top]
-                    set_default_options({sub1: 'Dynamically assigned'})
-                end
-            end
-            subject.run
-            subject.info[:top].should == 'nothing'
-            subject.info[:sub1].should == false
+            subject.config[:top].should == 'Command line magic'
+            subject.config[:sub1].should == 'Dynamically assigned'
         end
 
         it 'should have default options available from subcommands' do
             subject.args = ['sub1']
-            subject.set_default_options({top: 'A Symbol Key'})
+            subject.apply_config({top: 'A Symbol Key'})
             subject.run
             sub = subject.subactions.first
-            sub.info[:top].should == 'A Symbol Key'
+            sub.config[:top].should == 'A Symbol Key'
         end
 
         it 'should have default options overridden from commandline available from subcommands' do
             subject.args = ['--top', 'Bernard', 'sub1']
-            subject.set_default_options({top: 'A Symbol Key'})
+            subject.apply_config({top: 'A Symbol Key'})
             subject.run
             sub = subject.subactions.first
-            sub.info[:top].should == 'Bernard'
+            sub.config[:top].should == 'Bernard'
         end
 
         it 'should have default options work for subcommand options' do
-            subject.args = ['sub1', '--sub-one', 'Brian' ]
-            subject.set_default_options({sub_one: 'A Symbol Sub One'})
+            subject.args = ['sub1', '--sub-one']
+            subject.apply_config({sub_one: 'A Symbol Sub One'})
             subject.run
+            subject.config[:sub_one].should == true
             sub = subject.subactions.first
-            sub.info[:sub_one].should == 'Brian'
+            sub.config[:sub_one].should == true
+        end
+
+        it 'should allow the setting of config not in the options' do
+            subject.args = ['sub1', '--sub-one']
+            subject.apply_config({invisible: 'A Symbol Sub One'})
+            subject.run
+            subject.config[:invisible].should == 'A Symbol Sub One'
+            sub = subject.subactions.first
+            sub.config[:invisible].should == 'A Symbol Sub One'
         end
     end
 end